Move runtime host internals into runtimehost service
Interfaces and examples:

- Start runtime processes with runtimehost.StartPluginProcess(ctx, runtimehost.ProcessConfig{Command: command, Args: args, ProviderName: name}).

- Bind host services with runtimehost.StartHostServices([]runtimehost.HostService{{Name: "cache", EnvVar: "GESTALT_CACHE_SOCKET", Register: register}}, runtimehost.WithHostServicesProviderName(name)).

- Configure runtime lifecycle clients with runtimehost.ConfigureRuntimeProvider(ctx, process.Lifecycle(), kind, name, config).

- Register runtime log transport with runtimehost.RegisterRuntimeLogHostServer(srv, runtimeProviderName, appendLogs).

Providerhost remains as a compatibility layer for existing transports while the implementation now lives under services/runtimehost.
