+++ gestaltd/services/runtimehost/grpc_telemetry.go
@@ -1,6 +1,8 @@
-package providerhost
+package runtimehost
 
 import (
+	"strings"
+
 	"github.com/valon-technologies/gestalt/server/internal/metricutil"
 	"go.opentelemetry.io/contrib/instrumentation/google.golang.org/grpc/otelgrpc"
 )
@@ -9,8 +11,8 @@ func providerClientGRPCOptions(providerName string, telemetry metricutil.Telemet
 	return grpcTelemetryOptions(metricutil.RPCRoleProviderClient, providerName, "", telemetry)
 }
 
-func providerServerGRPCOptions(providerName string, telemetry metricutil.TelemetryProviders) []otelgrpc.Option {
-	return grpcTelemetryOptions(metricutil.RPCRoleProviderServer, providerName, "", telemetry)
+func hostServiceServerGRPCOptions(providerName string, service HostService, telemetry metricutil.TelemetryProviders) []otelgrpc.Option {
+	return grpcTelemetryOptions(metricutil.RPCRoleHostServiceServer, providerName, hostServiceMetricName(service), telemetry)
 }
 
 func grpcTelemetryOptions(role, providerName, hostService string, telemetry metricutil.TelemetryProviders) []otelgrpc.Option {
@@ -19,4 +21,17 @@ func grpcTelemetryOptions(role, providerName, hostService string, telemetry metr
 		ProviderName:    providerName,
 		HostServiceName: hostService,
 	})
+}
+
+func hostServiceMetricName(service HostService) string {
+	if service.Name != "" {
+		return strings.TrimSpace(service.Name)
+	}
+	envVar := strings.TrimSpace(service.EnvVar)
+	if envVar == "" {
+		return ""
+	}
+	envVar = strings.TrimPrefix(envVar, "GESTALT_")
+	envVar = strings.TrimSuffix(envVar, "_SOCKET")
+	return strings.ToLower(envVar)
 }
