+++ gestaltd/services/runtimehost/runtime_client_test.go
@@ -1,4 +1,4 @@
-package providerhost
+package runtimehost
 
 import (
 	"context"
