+++ gestaltd/services/runtimehost/runtime_log_env.go
@@ -1,0 +1,3 @@
+package runtimehost
+
+const DefaultRuntimeLogHostSocketEnv = "GESTALT_RUNTIME_LOG_SOCKET"
